% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.appendInfluence.R
\name{ds.AppendInfluence}
\alias{ds.AppendInfluence}
\title{generate prop scores server side}
\usage{
ds.AppendInfluence(
  df,
  influences,
  id_period_vector,
  column,
  newobj = NULL,
  datasources = NULL
)
}
\arguments{
\item{datasources}{connection}

\item{object}{table name}

\item{name}{saved new table name}

\item{pos}{distance measure}

\item{form}{formula used for glm}
}
\value{

}
\description{
generate prop scores server side
}
