% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildHelperDS.R
\name{ds.buildHelper}
\alias{ds.buildHelper}
\title{generate prop scores server side}
\usage{
ds.buildHelper(x, datasources = NULL)
}
\arguments{
\item{x}{just a helper}

\item{datasources}{connection}
}
\description{
generate prop scores server side
}
