% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.clusterInfluenceFunction.R
\name{ds.clusterInfluenceFunction}
\alias{ds.clusterInfluenceFunction}
\title{generate prop scores server side}
\usage{
ds.clusterInfluenceFunction(
  df,
  influence_matrix,
  clustervars,
  idname,
  newobj = NULL,
  datasources = NULL
)
}
\arguments{
\item{datasources}{connection}

\item{object}{table name}

\item{name}{saved new table name}

\item{pos}{distance measure}

\item{form}{formula used for glm}
}
\value{

}
\description{
generate prop scores server side
}
