% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.computeInfTreatDifference.R
\name{ds.computeInfTreatDifference}
\alias{ds.computeInfTreatDifference}
\title{generate prop scores server side}
\usage{
ds.computeInfTreatDifference(inf_1, inf_2, datasources = NULL)
}
\arguments{
\item{datasources}{connection}

\item{object}{table name}

\item{name}{saved new table name}

\item{pos}{distance measure}

\item{form}{formula used for glm}
}
\value{

}
\description{
generate prop scores server side
}
