% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.computeMatrixCrossproduct.R
\name{ds.computeMatrixCrossproduct}
\alias{ds.computeMatrixCrossproduct}
\title{computeMatrixCrossproduct}
\usage{
ds.computeMatrixCrossproduct(x, datasources = NULL)
}
\arguments{
\item{x}{A string of the matrix for which the cross product is to be computed.}

\item{datasources}{A specific Datashield data source to which the result should be assigned.}
}
\description{
The function computes the cross product (matrix transpose times matrix) of a matrix x which is stored on the server side and returns it to the client side.
}
