% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.computeOdds.R
\name{ds.computeOdds}
\alias{ds.computeOdds}
\title{generate prop scores server side}
\usage{
ds.computeOdds(name_propensities, name_C, newobj = NULL, datasources = NULL)
}
\arguments{
\item{datasources}{connection}

\item{object}{table name}

\item{name}{saved new table name}

\item{pos}{distance measure}

\item{form}{formula used for glm}
}
\value{

}
\description{
generate prop scores server side
}
