% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.createEmptyIdMatrix.R
\name{ds.createEmptyIdMatrix}
\alias{ds.createEmptyIdMatrix}
\title{createEmptyIdMatrix}
\usage{
ds.createEmptyIdMatrix(
  df,
  idname,
  n_columns,
  newobj = NULL,
  datasources = NULL
)
}
\arguments{
\item{df}{A string containing the name of the dataframe containing the unique identifier column.}

\item{idname}{The name of the column in the dataframe that serves as the unique identifier.}

\item{n_columns}{The number of columns for the empty matrix.}

\item{newobj}{The name of the object to which the result should be assigned.}

\item{datasources}{A specific Datashield data source to which the result should be assigned.}
}
\description{
The function creates an empty matrix with unique identifiers from a dataframe and a specified number of columns on the server side.umns for the matrix. The function also allows for the option to assign the result to a new object or a specific data source.
}
