% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.did.R
\name{ds.did}
\alias{ds.did}
\title{Compute did estimate
client side function ...}
\usage{
ds.did(
  data = NULL,
  yname = NULL,
  tname = NULL,
  idname = NULL,
  gname = NULL,
  t_periods = NULL,
  g_periods = NULL,
  xformla = NULL,
  control_group = "notyettreated",
  base_period = "varying",
  anticipation = 0,
  alpha = 0.05,
  bstrap = FALSE,
  biters = 1000,
  cband = TRUE,
  clustervars = NULL,
  est_method = "dr",
  maxit = 10000,
  seed_append = 12345,
  datasources = NULL,
  clear_console = FALSE
)
}
\arguments{
\item{df}{table name}

\item{pos}{distance measure}

\item{form}{formula}
}
\description{
Compute did estimate
client side function ...
}
