% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.generateNotYetTreated.R
\name{ds.generateNotYetTreated}
\alias{ds.generateNotYetTreated}
\title{generateNotYetTreated}
\usage{
ds.generateNotYetTreated(df, name_variable, t, g, newobj, datasources = NULL)
}
\arguments{
\item{df}{A string name of the dataframe containing the variables to be used.}

\item{name_variable}{The name of the variable to be used to identify treatment status.}

\item{t}{A time variable indicating when the treatment occurred.}

\item{g}{A group variable indicating to which group the individual belongs.}

\item{newobj}{The name of the object to which the result should be assigned on the server side.}

\item{datasources}{A specific Datashield data source to which the result should be assigned.}
}
\description{
The function generates a dataframe of individuals who have not yet been treated, given a dataframe, a variable name, a time variable and a group variable on the server side.
}
