% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.genProp.R
\name{ds.genProp}
\alias{ds.genProp}
\title{generate prop scores server side}
\usage{
ds.genProp(
  form,
  coefficients,
  object,
  constant_in_matrix = FALSE,
  newobj = "propscore",
  datasources = NULL,
  invlog = TRUE
)
}
\arguments{
\item{form}{formula used for glm}

\item{object}{table name}

\item{datasources}{connection}

\item{name}{saved new table name}

\item{pos}{distance measure}
}
\value{

}
\description{
generate prop scores server side
}
