% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.insertNumberToVector.R
\name{ds.insertNumberToVector}
\alias{ds.insertNumberToVector}
\title{generate prop scores server side}
\usage{
ds.insertNumberToVector(x, number, position, newobj = NULL, datasources = NULL)
}
\arguments{
\item{datasources}{connection}

\item{object}{table name}

\item{name}{saved new table name}

\item{pos}{distance measure}

\item{form}{formula used for glm}
}
\value{

}
\description{
generate prop scores server side
}
