% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.multiplierBootstrap.R
\name{ds.multiplierBootstrap}
\alias{ds.multiplierBootstrap}
\title{multiplierBootstrap}
\usage{
ds.multiplierBootstrap(matrix, n_iterations, datasources = NULL)
}
\arguments{
\item{matrix}{A string with the name of the matrix on which to perform the multiplier bootstrap.}

\item{n_iterations}{The number of iterations for the bootstrap.}

\item{datasources}{A specific Datashield data source to which the result should be assigned.}
}
\description{
The function performs a multiplier bootstrap on a given matrix on the server side and returns the results to the client side.
}
