% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.multiplyMatrixScalar.R
\name{ds.multiplyMatrixScalar}
\alias{ds.multiplyMatrixScalar}
\title{Multiply Matrix and Scalar}
\usage{
ds.multiplyMatrixScalar(matrix1, matrix2, newobj = NULL, datasources = NULL)
}
\arguments{
\item{newobj}{The name of the new object to assign the result of the multiplication to on the server side.}

\item{datasources}{A specific Datashield data source to which the result should be assigned.}

\item{matrix}{The matrix to be multiplied.  Must be a string with the name of the matrix on the server side.}

\item{scalar}{The scalar to be multiplied. Must be a single number.}
}
\description{
The function is used to multiply a matrix and a scalar and assign the result to a new object on the server side.
}
