% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.recode.R
\name{ds.recode}
\alias{ds.recode}
\title{ds.recode}
\usage{
ds.recode(x, replace_with, replace, newobj = NULL, datasources = NULL)
}
\arguments{
\item{x}{Vector in which values should be replaced}

\item{replace_with}{new value}

\item{replace}{Entry to be replaced}

\item{newobj}{The name of the object to which the result should be assigned.}

\item{datasources}{A specific Datashield data source to which the result should be assigned.}
}
\description{
The function creates a new object of the given vector and replaces certain entries of this vector
}
