% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.sendToServer.R
\name{ds.sendToServer}
\alias{ds.sendToServer}
\title{Send Object to Server}
\usage{
ds.sendToServer(x, newobj = NULL, datasources = NULL)
}
\arguments{
\item{x}{The number to be sent to the server.}

\item{newobj}{The name of the new object to assign the sent object to on the server.}

\item{datasources}{A specific Datashield data source to which the result should be assigned.}
}
\description{
The function is used to send an a single number to the server side.
}
