% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.subsetDf.R
\name{ds.subsetDf}
\alias{ds.subsetDf}
\title{Subset Dataframe}
\usage{
ds.subsetDf(df, name_variable, value, include_zero, newobj, datasources = NULL)
}
\arguments{
\item{df}{A string with the name of the dataframe from the server side to be subsetted.}

\item{name_variable}{The name of the variable to subset on.}

\item{value}{The value to subset the variable on.}

\item{include_zero}{Logical. Should zero values be included in the subset?}

\item{newobj}{The name of the new object to assign the subsetted dataframe to.}

\item{datasources}{A specific Datashield data source to which the result should be assigned.}
}
\description{
The function is used to subset a dataframe based on a specified variable and value. The subsetted dataframe is assigned to a new object on the server side.
}
