% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.did.R
\name{ds.did}
\alias{ds.did}
\title{clusterInfluenceFunction}
\usage{
ds.did(
  data = NULL,
  yname = NULL,
  tname = NULL,
  idname = NULL,
  gname = NULL,
  t_periods = NULL,
  g_periods = NULL,
  xformla = NULL,
  control_group = "notyettreated",
  base_period = "varying",
  anticipation = 0,
  alpha = 0.05,
  bstrap = FALSE,
  biters = 1000,
  cband = TRUE,
  clustervars = NULL,
  est_method = "dr",
  maxit = 10000,
  seed_append = 12345,
  datasources = NULL,
  clear_console = FALSE
)
}
\arguments{
\item{clustervars}{A string name of the vector of variables that make up the cluster for standard errors. Only has an influence if bstrap is true.}

\item{datasources}{A specific Datashield data source to which the result should be assigned.}

\item{df}{A string name of the dataframe containing the data on the server side.}
}
\value{
aaa
}
\description{
The function calculates the influence of a cluster of variables on a target variable using an influence matrix on the server side.lows for the option to assign the result to a new object or a specific data source.
}
