% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.did.R
\name{ds.did}
\alias{ds.did}
\title{Compute did estimate
client side function ...}
\usage{
ds.did(
  data,
  yname,
  tname,
  idname,
  gname,
  t_periods,
  g_periods,
  xformla = NULL,
  panel = TRUE,
  allow_unbalanced_panel = FALSE,
  control_group = "notyettreated",
  base_period = "varying",
  anticipation = 0,
  alpha = 0.05,
  bstrap = FALSE,
  biters = 1000,
  cband = TRUE,
  clustervars = NULL,
  est_method = "dr",
  maxit = 10000,
  datasources = NULL
)
}
\arguments{
\item{df}{table name}

\item{pos}{distance measure}

\item{form}{formula}
}
\value{

}
\description{
Compute did estimate
client side function ...
}
