% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.enoughIndividuals.R
\name{ds.enoughIndividuals}
\alias{ds.enoughIndividuals}
\title{generate prop scores server side}
\usage{
ds.enoughIndividuals(df, colname, value, datasources = NULL)
}
\arguments{
\item{datasources}{connection}

\item{minimal_number}{}
}
\description{
generate prop scores server side
}
