% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.generateNotYetTreated.R
\name{ds.generateNotYetTreated}
\alias{ds.generateNotYetTreated}
\title{match it function (pooled)}
\usage{
ds.generateNotYetTreated(df, name_variable, t, g, newobj, datasources = NULL)
}
\arguments{
\item{datasources}{connection}

\item{object}{table name}

\item{name}{saved new table name}

\item{pos}{distance measure}

\item{form}{formula used for glm}
}
\value{

}
\description{
match it function (pooled)
}
