% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.genProp.R
\name{ds.genProp}
\alias{ds.genProp}
\title{genProp}
\usage{
ds.genProp(
  form,
  coefficients,
  object,
  constant_in_matrix = FALSE,
  newobj = "propscore",
  datasources = NULL,
  invlog = TRUE
)
}
\arguments{
\item{form}{A string describing the formula for generating the propensity scores}

\item{coefficients}{A vector of coefficients for the formula}

\item{object}{The ds.glm object to be used for generating the scores}

\item{constant_in_matrix}{A logical indicating whether to include a constant in the matrix}

\item{newobj}{The name of the object to which the result should be assigned on the server side.}

\item{datasources}{A specific Datashield data source to which the result should be assigned.}

\item{invlog}{A logical indicating whether to use the inverse logit transformation.}
}
\description{
The function generates propensity scores for a given formula, coefficients, and ds.glm object on the server side.
}
