% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.multiplierBootstrap.R
\name{ds.multiplierBootstrap}
\alias{ds.multiplierBootstrap}
\title{generate prop scores server side}
\usage{
ds.multiplierBootstrap(matrix, n_iterations, datasources = NULL)
}
\arguments{
\item{datasources}{connection}

\item{object}{table name}

\item{name}{saved new table name}

\item{pos}{distance measure}

\item{form}{formula used for glm}
}
\value{

}
\description{
generate prop scores server side
}
