% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.multiplyMatrixMatrix.R
\name{ds.multiplyMatrixMatrix}
\alias{ds.multiplyMatrixMatrix}
\title{generate prop scores server side}
\usage{
ds.multiplyMatrixMatrix(
  matrix1,
  matrix2,
  nrow1 = NULL,
  ncol1 = NULL,
  nrow2 = NULL,
  ncol2 = NULL,
  newobj = NULL,
  datasources = NULL
)
}
\arguments{
\item{datasources}{connection}

\item{object}{table name}

\item{name}{saved new table name}

\item{pos}{distance measure}

\item{form}{formula used for glm}
}
\value{

}
\description{
generate prop scores server side
}
