% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.sendToServer.R
\name{ds.sendToServer}
\alias{ds.sendToServer}
\title{generate prop scores server side}
\usage{
ds.sendToServer(x, newobj = NULL, datasources = NULL)
}
\arguments{
\item{datasources}{connection}

\item{object}{table name}

\item{name}{saved new table name}

\item{pos}{distance measure}

\item{form}{formula used for glm}
}
\description{
generate prop scores server side
}
