% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.subsetDf.R
\name{ds.subsetDf}
\alias{ds.subsetDf}
\title{match it function (pooled)}
\usage{
ds.subsetDf(df, name_variable, value, include_zero, newobj, datasources = NULL)
}
\arguments{
\item{datasources}{connection}

\item{object}{table name}

\item{name}{saved new table name}

\item{pos}{distance measure}

\item{form}{formula used for glm}
}
\value{

}
\description{
match it function (pooled)
}
